sincronização 0 falhas

+++ src/sync-threads.c
@@ -5,7 +5,7 @@
 #include "sync-threads.h"
 
 #define EVENT_SIZE (sizeof(struct inotify_event))
-#define EVENT_BUF_LEN (1024 * (EVENT_SIZE))
+#define EVENT_BUF_LEN (1024 * (EVENT_SIZE + MAXNAME))
 
 void *answer_pending(void* user) {
 	UserInfo *user_info = (UserInfo*) user;
@@ -55,20 +55,23 @@ void *watcher(void* user) {
 	strcpy(watch_path, (char*) user_info->folder);
 
 	int fd, wd;
-	int length, i = 0;
-	char buffer[EVENT_BUF_LEN];
 
 	fd = inotify_init();
 	if(fd < 0) {
 		printf("Error inotify_init\n");		
 	}
 	
-	wd = inotify_add_watch(fd, watch_path, IN_CLOSE_WRITE | IN_DELETE | IN_CREATE |  IN_DELETE_SELF | IN_MOVED_FROM | IN_MOVED_TO);
+	wd = inotify_add_watch(fd, watch_path, IN_CLOSE_WRITE | IN_MOVED_FROM | IN_MOVED_TO | IN_CREATE | IN_DELETE | IN_DELETE_SELF);
 
 	char path[MAXNAME];
 	int thread_running = TRUE;
+	int changes;
 
 	while(thread_running) {
+		int length, i = 0;
+		char buffer[EVENT_BUF_LEN];
+		changes = 0;
+
 		length = read(fd, buffer, EVENT_BUF_LEN); 
 		pthread_mutex_lock(&(user_info->lock_server_comm));
 
@@ -80,31 +83,44 @@ void *watcher(void* user) {
 				struct inotify_event* event = (struct inotify_event *) &buffer[i];
 
 				if (event->len) {
-					sprintf(path, "%s/%s", watch_path, event->name);
+					sprintf(path, "%s/%s", user_info->folder, event->name);
 
-					if (event->mask & (IN_CLOSE_WRITE | IN_CREATE | IN_MOVED_TO)) {
-						if(IN_CLOSE_WRITE) printf("\nIN_CLOSE_WRITE %s", event->name);
-						if(IN_CREATE) printf("\nIN_CREATE %s", event->name);
-						if(IN_MOVED_TO) printf("\nIN_MOVED_TO %s", event->name);
+					if (event->mask & (IN_CLOSE_WRITE | IN_MOVED_TO | IN_DELETE | IN_DELETE_SELF | IN_MOVED_FROM)) {
+						/*if(IN_CLOSE_WRITE) printf("\nIN_CLOSE_WRITE %s", path);
+						if(IN_MOVED_TO) printf("\nIN_MOVED_TO %s", path);
+						if(IN_DELETE) printf("\nIN_DELETE %s", path);
+						if(IN_DELETE_SELF) printf("\nIN_DELETE_SELF %s", path);
+						if(IN_MOVED_FROM) printf("\nIN_MOVED_FROM %s", path);*/
+						if(event->name[0] != '.')
+							changes++;
+					}
+						/*if(IN_CLOSE_WRITE) printf("\nIN_CLOSE_WRITE %s", path);
+						if(IN_MOVED_TO) printf("\nIN_MOVED_TO %s", path);
 						if (check_dir(path) && (event->name[0] != '.')) {
 							printf("\nRequest upload: %s to user_%s\n", event->name, user_info->id);
 
 							send_file(path, user_info, user_info->msg_id, TRUE);
 						}
-					} else if (event->mask & (IN_DELETE | IN_DELETE_SELF | IN_MOVED_FROM)) {
-						if(IN_DELETE) printf("\nIN_DELETE %s", event->name);
-						if(IN_DELETE_SELF) printf("\nIN_DELETE_SELF %s", event->name);
-						if(IN_MOVED_FROM) printf("\nIN_MOVED_FROM %s", event->name);
-						if (event->name[0] != '.') {
+					} 
+					if (event->mask & (IN_DELETE | IN_DELETE_SELF | IN_MOVED_FROM)) {
+						if(IN_DELETE) printf("\nIN_DELETE %s", path);
+						if(IN_DELETE_SELF) printf("\nIN_DELETE_SELF %s", path);
+						if(IN_MOVED_FROM) printf("\nIN_MOVED_FROM %s", path);
+						if (!check_dir(path)  && event->name[0] != '.') {
 							printf("\nRequest delete: %s to user_%s\n", event->name, user_info->id);
 
 							delete_file(event->name, user_info, user_info->msg_id, TRUE);;
 						}
-					}
+					}*/					
 				}
 
-			i += EVENT_SIZE + event->len;
+				i += EVENT_SIZE + event->len;
       		}
+			if(changes > 0) {
+				printf("\nRequesting server update");
+				update_server(user_info, user_info->msg_id);
+			}
+
 		}
 		pthread_mutex_unlock(&(user_info->lock_server_comm));;
 		usleep(10000000);
